requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.9
    # via kombu
appdirs==1.4.4
    # via requests-cache
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via
    #   cattrs
    #   pytest
    #   requests-cache
billiard==3.6.4.0
    # via celery
black==21.9b0
    # via -r requirements.in
cattrs==1.8.0
    # via requests-cache
celery==5.2.3
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.6
    # via requests
click==8.0.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coverage==5.5
    # via pytest-cov
deprecated==1.2.13
    # via redis
django==4.0.4
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-modeltranslation
    #   django-munigeo
    #   django-polymorphic
    #   django-timezone-field
    #   djangorestframework
django-celery-beat @ git+https://github.com/celery/django-celery-beat.git
    # via -r requirements.in
django-celery-results==2.2.0
    # via -r requirements.in
django-cors-headers==3.9.0
    # via -r requirements.in
django-environ==0.7.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-mptt
django-modeltranslation==0.17.3
    # via
    #   -r requirements.in
    #   django-munigeo
django-mptt==0.13.4
    # via
    #   -r requirements.in
    #   django-munigeo
django-munigeo @ git+https://github.com/City-of-Helsinki/django-munigeo@v0.2.69
    # via -r requirements.in
django-polymorphic==3.1.0
    # via -r requirements.in
django-timezone-field==4.2.3
    # via django-celery-beat
djangorestframework==3.12.4
    # via -r requirements.in
djangorestframework-jsonp==1.0.2
    # via -r requirements.in
flake8==3.9.2
    # via
    #   -r requirements.in
    #   flake8-polyfill
    #   pep8-naming
flake8-polyfill==1.0.2
    # via pep8-naming
idna==3.2
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.9.3
    # via -r requirements.in
jedi==0.18.0
    # via -r requirements.in
kombu==5.2.3
    # via celery
libvoikko==4.3
    # via -r requirements.in
lxml==4.8.0
    # via
    #   -r requirements.in
    #   pykml
markdown==3.3.4
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
numpy==1.21.2
    # via pandas
packaging==21.0
    # via
    #   pytest
    #   redis
pandas==1.3.3
    # via -r requirements.in
parso==0.8.2
    # via
    #   -r requirements.in
    #   jedi
pathspec==0.9.0
    # via black
pep517==0.11.0
    # via pip-tools
pep8-naming==0.12.1
    # via -r requirements.in
pip-tools==6.3.0
    # via -r requirements.in
platformdirs==2.4.0
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.26
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements.in
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pykml==0.2.0
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==2.12.1
    # via -r requirements.in
pytest-django==4.4.0
    # via -r requirements.in
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
    #   python-crontab
pytz==2021.3
    # via
    #   -r requirements.in
    #   celery
    #   django-timezone-field
    #   pandas
pyyaml==5.4.1
    # via django-munigeo
raven==6.10.0
    # via -r requirements.in
redis==4.1.3
    # via -r requirements.in
regex==2021.9.24
    # via black
requests==2.26.0
    # via
    #   -r requirements.in
    #   django-munigeo
    #   requests-cache
    #   requests-mock
requests-cache==0.8.1
    # via -r requirements.in
requests-mock==1.9.3
    # via -r requirements.in
shapely==1.8.0
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   django-modeltranslation
    #   django-munigeo
    #   python-dateutil
    #   requests-mock
    #   url-normalize
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
tomli==1.2.1
    # via
    #   black
    #   pep517
tqdm==4.62.3
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via black
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.7
    # via
    #   requests
    #   requests-cache
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.0
    # via pip-tools
whitenoise==5.3.0
    # via -r requirements.in
wrapt==1.13.3
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
